# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API using Gin framework",
        "title": "API",
        "contact": {}
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/students": {
            "get": {
                "description": "Return list of students.",
                "summary": "Get All student.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "obejct"
                        }
                    }
                }
            },
            "post": {
                "description": "Save a student data in Db.",
                "produces": [
                    "application/json"
                ],
                "summary": "Create a student",
                "parameters": [
                    {
                        "description": "create a student",
                        "name": "student",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.HttpResponse"
                        }
                    }
                }
            }
        },
        "/students/{studentId}": {
            "get": {
                "description": "Return the tahs whoes studentId valu mathes id.",
                "produces": [
                    "application/json"
                ],
                "summary": "Get Single student by id.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "update student by id",
                        "name": "studentId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.HttpResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Remove student data by id.",
                "produces": [
                    "application/json"
                ],
                "summary": "Delete a student",
                "parameters": [
                    {
                        "type": "string",
                        "description": "delete a student by id",
                        "name": "studentId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.HttpResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "Update student data.",
                "produces": [
                    "application/json"
                ],
                "summary": "Update a student",
                "parameters": [
                    {
                        "type": "string",
                        "description": "update student by id",
                        "name": "studentId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update a student",
                        "name": "student",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.HttpResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "request.CreateRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "maxLength": 200,
                    "minLength": 1
                }
            }
        },
        "response.HttpResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "status": {
                    "type": "string"
                }
            }
        }
    }
}